# Proposal : Optimize locking for create container request