--- a/scheduler/filter/affinity.go
+++ b/scheduler/filter/affinity.go
@@ -36,6 +36,9 @@ func (f *AffinityFilter) Filter(config *dockerclient.ContainerConfig, nodes []*n
 				for _, container := range node.Containers {
 					containers = append(containers, container.Id, strings.TrimPrefix(container.Names[0], "/"))
 				}
+				for _, container := range node.ScheduledList("container") {
+					containers = append(containers, container)
+				}
 				if affinity.Match(containers...) {
 					candidates = append(candidates, node)
 				}
@@ -47,6 +50,9 @@ func (f *AffinityFilter) Filter(config *dockerclient.ContainerConfig, nodes []*n
 					for _, tag := range image.RepoTags {
 						images = append(images, strings.Split(tag, ":")[0])
 					}
+				}
+				for _, image := range node.ScheduledList("images") {
+					images = append(images, image)
 				}
 				if affinity.Match(images...) {
 					candidates = append(candidates, node)