@vieux 2015-10-14
I believe a similar concept was added in https://github.com/docker/swarm/pull/1261